requirements/requirements.txt
# Django
Django==4.1.2

# Misc
loguru==0.6.0

# Third party
celery==5.2.7
dj-database-url==0.5.0
django-redis==5.2.0
django-celery-results==2.4.0
django-crispy-forms==1.14.0
djangorestframework>=3.13.1,!=3.14.0
djangorestframework-simplejwt==5.2.1
drf-spectacular-sidecar==2022.9.1
drf-spectacular==0.23.1
Pillow==9.2.0
requests==2.28.1
lxml==4.9.1
beautifulsoup4>=4.11.1
